CuratedData class initialization improved
The most basic functions of moving average, deviation and RSI have now an additional argument to decide whether the results must be appended in the data-frame or returned. In case append is passed as True (default is False, i.e, by default it returns the result) the parameters dictionary in the class is updated.