Improvements in autocorrelation method
There are 3 methods involving auto-correlation functions now. The first compute in a given period of the data-frame defined by pandas.Timestamp. The second, compute the autocorrelation in the last few data points given by a window. The third, compute a moving autocorrelation, which provide a option to append as a feature (This last one is usually high demanding).