Matrices of autocorrelation
New functions to compute autocorrelations automatically for a set of inputs in a list and assemble the result as a matrix (actually pandas dataframe). The method 'autocorr_period_matrix' assemble a matrix with several start and end times enclosing windows in the dataframe. The method 'autocorr_tail_matrix' proceeds in the same way but the dataframe is indexed by the windows and shifts size, where the windows are the number of the most recent data points (dataframe tail)