Close #14 rawdata method improved
Introduced two numba function to assist in accumulated sums computation, avoiding python loops. New private method assemble the (new) resulting dataframe, procedure similar for tick, volume and money bars. Function to convert dataframe sample time interval could not be cast in a suitable format to use numba, due to unpredictable time gaps in dataframe caused either due to market circuit breaker or low deal rate, and as such, we cannot count on minute-to-minute data. All methods to re-format the dataframe now provide an start and stop indexing instead to compulsorily apply to the entirde dataframe, which is still the default case if no parameters are given.