Improvements and new features in labels.py
The improvements consist in avoiding excessive use of conditionals and get rid of lists and append. The new features consist in using allowing different types of investiment horizon, as timedelta and cummulative sum of some field as volume/tickvol/money. As such, the entire dataframe is required in contrast to just the close prices. This also provide a way to generalize the implementation to include some spread in the prices to simulate a real world trade system with the high and low values.