More elegant and robust implementation of return series of a portfolio. Some bugs fixed
The implementation of the return series now work for a selected symbol and provide a cleaner implementation also to the overall net result of the porfolion. Found a bug in PortfolioRecord class since lists of deals were returned and were prone to have unpredictable behavior if some deal was modified outside de class. Bug corrected.